Refactor menu.js into UI modules and fix clipboard/invite/auth bugs
Split the monolithic menu into ui/* and input/keys composition modules,
coordinate khet editing state via khet/editingState.js, remove the WASM
password gate and unsafe-eval CSP, and fix username/peer-id copy,
undefined generateInvitation, saveBtn early-return abort, hardcoded
canister share links, and related client hardening.
