+++ src/otherland_client/.ic-assets.json5
@@ -2,7 +2,7 @@
   {
     "match": "*.html",
     "headers": {
-      "Content-Security-Policy": "default-src 'self'; script-src 'self' 'unsafe-eval' https://cdn.jsdelivr.net; object-src 'none'; connect-src 'self' 0.peerjs.com wss://0.peerjs.com http://localhost:4943 https://icp0.io https://cdn.jsdelivr.net blob:; font-src 'self' data:",
+      "Content-Security-Policy": "default-src 'self'; script-src 'self' https://cdn.jsdelivr.net; object-src 'none'; connect-src 'self' 0.peerjs.com wss://0.peerjs.com http://localhost:4943 http://*.localhost:4943 https://icp0.io https://*.icp0.io https://icp-api.io https://cdn.jsdelivr.net blob:; font-src 'self' data:",
       "Permissions-Policy": "xr-spatial-tracking=(self)"
     },
     "security_policy": "hardened"
@@ -10,14 +10,14 @@
   {
     "match": "*.js",
     "headers": {
-      "Content-Security-Policy": "default-src 'self'; script-src 'self'; object-src 'none'; connect-src 'self' https://icp0.io;"
+      "Content-Security-Policy": "default-src 'self'; script-src 'self'; object-src 'none'; connect-src 'self' https://icp0.io https://*.icp0.io https://icp-api.io http://*.localhost:4943;"
     },
     "security_policy": "hardened"
   },
   {
     "match": "*.css",
     "headers": {
-      "Content-Security-Policy": "default-src 'self'; script-src 'self'; object-src 'none'; connect-src 'self' https://icp0.io; font-src 'self' data:"
+      "Content-Security-Policy": "default-src 'self'; script-src 'self'; object-src 'none'; connect-src 'self' https://icp0.io https://*.icp0.io https://icp-api.io; font-src 'self' data:"
     },
     "security_policy": "hardened"
   },
