+++ src/otherland_client/src/khet/codeExecutor.js
@@ -1,0 +1,53 @@
+import * as esprima from 'esprima';
+
+// **Khet Code Interpreter**
+// Simple interpreter for Khet code using Esprima to parse and validate expressions
+export function createKhetCodeExecutor(code, object) {
+    try {
+        const ast = esprima.parseScript(code); // Parse the code into an Abstract Syntax Tree (AST)
+        if (ast.body.length !== 1 || ast.body[0].type !== 'ExpressionStatement') {
+            console.warn(`Khet code must be a single expression: ${code}`);
+            return () => {};
+        }
+        const expr = ast.body[0].expression;
+        if (expr.type !== 'AssignmentExpression' || !['=', '+=', '-=', '*=', '/='].includes(expr.operator)) {
+            console.warn(`Unsupported operation in Khet code: ${code}`);
+            return () => {};
+        }
+        const left = expr.left;
+        if (left.type !== 'MemberExpression' || 
+            left.object.type !== 'MemberExpression' || 
+            left.object.object.type !== 'Identifier' || 
+            left.object.object.name !== 'object') {
+            console.warn(`Khet code must assign to object.property.axis: ${code}`);
+            return () => {};
+        }
+        const property = left.object.property.name;
+        const axis = left.property.name;
+        const allowedProperties = ['rotation', 'position', 'scale'];
+        if (!allowedProperties.includes(property) || !['x', 'y', 'z'].includes(axis)) {
+            console.warn(`Invalid property or axis in Khet code: ${code}`);
+            return () => {};
+        }
+        const right = expr.right;
+        if (right.type !== 'Literal' || typeof right.value !== 'number') {
+            console.warn(`Khet code right-hand side must be a number: ${code}`);
+            return () => {};
+        }
+        const value = right.value;
+        const operator = expr.operator;
+        // Return a function that executes the validated assignment operation
+        return () => {
+            switch (operator) {
+                case '=': object[property][axis] = value; break;
+                case '+=': object[property][axis] += value; break;
+                case '-=': object[property][axis] -= value; break;
+                case '*=': object[property][axis] *= value; break;
+                case '/=': object[property][axis] /= value; break;
+            }
+        };
+    } catch (error) {
+        console.error(`Error parsing Khet code: ${code}`, error);
+        return () => {};
+    }
+}
