+++ src/otherland_client/src/khet/editingState.js
@@ -1,0 +1,10 @@
+/** Mutable khet editor selection state (shared by khet table + asset UI). */
+export let currentEditingKhetId = null;
+
+export function setCurrentEditingKhetId(id) {
+    currentEditingKhetId = id;
+}
+
+export function clearCurrentEditingKhetId() {
+    currentEditingKhetId = null;
+}
