+++ src/otherland_client/src/utils/sha256.js
@@ -1,0 +1,7 @@
+// Compute SHA-256 hash of a Uint8Array
+export async function computeSHA256(data) {
+    const buffer = await crypto.subtle.digest('SHA-256', data);
+    return Array.from(new Uint8Array(buffer))
+        .map(b => b.toString(16).padStart(2, '0'))
+        .join('');
+}
