+++ src/otherland_client/vite.config.js
@@ -54,6 +54,6 @@ export default defineConfig({
         ),
       },
     ],
-    dedupe: ['@dfinity/agent'],
+    dedupe: ['@icp-sdk/core'],
   },
 });
